// frontend/src/api/generated/models/property-type.ts
/* tslint:disable */
/* eslint-disable */
/**
 * chainlink-api API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const PropertyType = {
    Text: 'TEXT',
    Date: 'DATE',
    Select: 'SELECT',
    MultiSelect: 'MULTI_SELECT',
    Boolean: 'BOOLEAN',
    Number: 'NUMBER'
} as const;
export type PropertyType = typeof PropertyType[keyof typeof PropertyType];


export function instanceOfPropertyType(value: any): boolean {
    for (const key in PropertyType) {
        if (Object.prototype.hasOwnProperty.call(PropertyType, key)) {
            if (PropertyType[key as keyof typeof PropertyType] === value) {
                return true;
            }
        }
    }
    return false;
}

export function PropertyTypeFromJSON(json: any): PropertyType {
    return PropertyTypeFromJSONTyped(json, false);
}

export function PropertyTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): PropertyType {
    return json as PropertyType;
}

export function PropertyTypeToJSON(value?: PropertyType | null): any {
    return value as any;
}

export function PropertyTypeToJSONTyped(value: any, ignoreDiscriminator: boolean): PropertyType {
    return value as PropertyType;
}